moved utils to helpers folder and added logging of event and response
