+++ product-service/package.json
@@ -21,6 +21,7 @@
     "@middy/http-cors": "^3.1.1",
     "@middy/http-error-handler": "^3.1.1",
     "@middy/http-response-serializer": "^3.1.1",
+    "@middy/input-output-logger": "^3.2.0",
     "@nestjs/common": "^9.0.0",
     "@nestjs/core": "^9.0.0",
     "http-errors": "^2.0.0",
